requirements.txt
# Core Web Framework
streamlit>=1.28.0,<2.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# ML/AI
openai>=1.40.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
joblib>=1.3.0,<2.0.0

# Configuration & Environment
python-dotenv>=1.0.0,<2.0.0

# Optional: Development & Notebooks
jupyter>=1.0.0,<2.0.0
ipykernel>=6.0.0,<7.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
